Address coderabbit review
