+++ backend/src/apps/owasp/models/crp/certificate.py
@@ -49,6 +49,21 @@ class Certificate(TimestampedModel):
                 name="unique_active_cert_per_tier",
                 violation_error_message="Cannot have multiple active certificates for same tier",
             ),
+            models.CheckConstraint(
+                condition=(
+                    (~Q(tier="") & Q(tier__isnull=False))
+                    | (
+                        ~Q(title="")
+                        & Q(title__isnull=False)
+                        & (Q(project__isnull=False) | Q(chapter__isnull=False))
+                    )
+                ),
+                name="valid_certificate_type",
+                violation_error_message=(
+                    "Certificate must be either a valid tier certificate or a generic certificate"
+                    " with a title and associated project/chapter."
+                ),
+            ),
         ]
 
     id = models.CharField(
